refactor: Room 관련 Service 분리
