+++ src/main/resources/static/docs/notification.html
@@ -487,7 +487,7 @@ Content-Length: 64
 <div id="footer">
 <div id="footer-text">
 Version 0.0.1-SNAPSHOT<br>
-Last updated 2023-11-10 21:12:31 +0900
+Last updated 2023-11-12 16:01:40 +0900
 </div>
 </div>
 </body>
